third_party/WebKit/LayoutTests/fast/dom/inline-event-attributes-moved-expected.txt

Tests that we have the expected form in scope

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS expected is "undefined"
PASS successfullyParsed is true

TEST COMPLETE
